grunt 4.0 and JSDocs
Reset grunt to run 4.0 and added JSDocs into the build cycle.
There is also an option for YuiDocs, but so far is doesn't seem
as immediately good -- left for now so I can try it again once the
docs have been build up. Also cleaned out some of the scratch test files